Can now serve also css and javascript also finished up cleaning memmory

+++ Headers/Utils/memmory.h
@@ -6,6 +6,6 @@
 void *allocate(size_t size);
 void deallocate(void *block);
 void *reallocate(void *block, size_t size);
-int getAllocations(void);
+size_t getAllocations(void);
 
 #endif
